feat(nav): added navigation to app
