+++ ios/TipTap/Images.xcassets/AppIcon.appiconset/Contents.json
@@ -1,38 +1,181 @@
 {
-  "images" : [
-    {
-      "idiom" : "iphone",
-      "size" : "29x29",
-      "scale" : "2x"
+    "images": [
+        {
+            "size": "20x20",
+            "idiom": "iphone",
+            "filename": "Icon-App-20x20@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "20x20",
+            "idiom": "iphone",
+            "filename": "Icon-App-20x20@3x.png",
+            "scale": "3x"
+        },
+        {
+            "size": "29x29",
+            "idiom": "iphone",
+            "filename": "Icon-App-29x29@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "29x29",
+            "idiom": "iphone",
+            "filename": "Icon-App-29x29@3x.png",
+            "scale": "3x"
+        },
+        {
+            "size": "40x40",
+            "idiom": "iphone",
+            "filename": "Icon-App-40x40@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "40x40",
+            "idiom": "iphone",
+            "filename": "Icon-App-40x40@3x.png",
+            "scale": "3x"
+        },
+        {
+            "size": "60x60",
+            "idiom": "iphone",
+            "filename": "Icon-App-60x60@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "60x60",
+            "idiom": "iphone",
+            "filename": "Icon-App-60x60@3x.png",
+            "scale": "3x"
+        },
+        {
+            "size": "20x20",
+            "idiom": "ipad",
+            "filename": "Icon-App-20x20@1x.png",
+            "scale": "1x"
+        },
+        {
+            "size": "20x20",
+            "idiom": "ipad",
+            "filename": "Icon-App-20x20@2x-1.png",
+            "scale": "2x"
+        },
+        {
+            "size": "29x29",
+            "idiom": "ipad",
+            "filename": "Icon-App-29x29@1x.png",
+            "scale": "1x"
+        },
+        {
+            "size": "29x29",
+            "idiom": "ipad",
+            "filename": "Icon-App-29x29@2x-1.png",
+            "scale": "2x"
+        },
+        {
+            "size": "40x40",
+            "idiom": "ipad",
+            "filename": "Icon-App-40x40@1x.png",
+            "scale": "1x"
+        },
+        {
+            "size": "40x40",
+            "idiom": "ipad",
+            "filename": "Icon-App-40x40@2x-1.png",
+            "scale": "2x"
+        },
+        {
+            "size": "76x76",
+            "idiom": "ipad",
+            "filename": "Icon-App-76x76@1x.png",
+            "scale": "1x"
+        },
+        {
+            "size": "76x76",
+            "idiom": "ipad",
+            "filename": "Icon-App-76x76@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "83.5x83.5",
+            "idiom": "ipad",
+            "filename": "Icon-App-83.5x83.5@2x.png",
+            "scale": "2x"
+        },
+        {
+            "size": "1024x1024",
+            "idiom": "ios-marketing",
+            "filename": "ItunesArtwork@2x.png",
+            "scale": "1x"
+        },
+        {
+            "size": "24x24",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-24@2x.png",
+            "role": "notificationCenter",
+            "subtype": "38mm"
+        },
+        {
+            "size": "27.5x27.5",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-27.5@2x.png",
+            "role": "notificationCenter",
+            "subtype": "42mm"
+        },
+        {
+            "size": "29x29",
+            "idiom": "watch",
+            "filename": "Icon-29@2x.png",
+            "role": "companionSettings",
+            "scale": "2x"
+        },
+        {
+            "size": "29x29",
+            "idiom": "watch",
+            "filename": "Icon-29@3x.png",
+            "role": "companionSettings",
+            "scale": "3x"
+        },
+        {
+            "size": "40x40",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-40@2x.png",
+            "role": "appLauncher",
+            "subtype": "38mm"
+        },
+        {
+            "size": "44x44",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-44@2x.png",
+            "role": "longLook",
+            "subtype": "42mm"
+        },
+        {
+            "size": "86x86",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-86@2x.png",
+            "role": "quickLook",
+            "subtype": "38mm"
+        },
+        {
+            "size": "98x98",
+            "idiom": "watch",
+            "scale": "2x",
+            "filename": "Icon-98@2x.png",
+            "role": "quickLook",
+            "subtype": "42mm"
+        }
+    ],
+    "info": {
+        "version": 1,
+        "author": "fanstudio"
     },
-    {
-      "idiom" : "iphone",
-      "size" : "29x29",
-      "scale" : "3x"
-    },
-    {
-      "idiom" : "iphone",
-      "size" : "40x40",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "iphone",
-      "size" : "40x40",
-      "scale" : "3x"
-    },
-    {
-      "idiom" : "iphone",
-      "size" : "60x60",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "iphone",
-      "size" : "60x60",
-      "scale" : "3x"
+    "properties": {
+        "pre-rendered": true
     }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
 }
