+++ package.json
@@ -19,7 +19,8 @@
   },
   "dependencies": {
     "react": "16.2.0",
-    "react-native": "0.53.3"
+    "react-native": "0.53.3",
+    "react-navigation": "^1.2.1"
   },
   "devDependencies": {
     "babel-jest": "22.4.1",
