+++ src/containers/App/App.js
@@ -4,33 +4,23 @@
  * @flow
  */
 
-import React, { Component } from 'react';
-import { Platform, Text, View } from 'react-native';
+import React from 'react';
+import { StackNavigator } from 'react-navigation';
 
-import styles from './App.styles';
+// Screens
+import Home from './../Home/Home';
 
-const instructions = Platform.select({
-  ios: 'Press Cmd+R to reload,\n' +
-    'Cmd+D or shake for dev menu',
-  android: 'Double tap R on your keyboard to reload,\n' +
-    'Shake or press menu button for dev menu',
-});
+// Routing
+const RootStack = StackNavigator(
+  // Route Map
+  {
+    Home: { screen: Home },
+  },
+  // Stack Config
+  {
+    initialRouteName: 'Home',
+  },
+);
 
-type Props = {};
-export default class App extends Component<Props> {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to TipTap!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit App.js in src/containers/App/
-        </Text>
-        <Text style={styles.instructions}>
-          {instructions}
-        </Text>
-      </View>
-    );
-  }
-}
+// Export Root Stack
+export default () => <RootStack />;
