+++ src/containers/Home/Home.js
@@ -1,0 +1,41 @@
+/**
+ * TipTap - Tip Calculator
+ * https://github.com/GoodHombres/TipTap
+ * @flow
+ */
+
+import React, { Component } from 'react';
+import { Platform, Text, View } from 'react-native';
+
+import styles from './Home.styles';
+
+const instructions = Platform.select({
+  ios: 'Press Cmd+R to reload,\n' +
+    'Cmd+D or shake for dev menu',
+  android: 'Double tap R on your keyboard to reload,\n' +
+    'Shake or press menu button for dev menu',
+});
+
+type Props = {};
+export default class Home extends Component<Props> {
+  // Hide navigation header
+  static navigationOptions = {
+    header: null,
+  };
+
+  render() {
+    return (
+      <View style={styles.container}>
+        <Text style={styles.welcome}>
+          Welcome to TipTap!
+        </Text>
+        <Text style={styles.instructions}>
+          To get started, edit Home.js in src/containers/Home/
+        </Text>
+        <Text style={styles.instructions}>
+          {instructions}
+        </Text>
+      </View>
+    );
+  }
+}
